requirements.txt
# Echovita Pipeline — Dependencies
# Python 3.11+

# Web Scraping
scrapy==2.14.1
itemadapter==0.10.0

# Data Processing
duckdb==1.5.0
pandas==2.2.3
numpy==2.2.4

# Dashboard
streamlit==1.42.0

# HTTP
requests==2.32.3

# Airflow — core only
apache-airflow==2.10.4

# Utils
python-dateutil==2.9.0

# Testing
pytest==8.3.5
